requirements.txt
paho-mqtt~=1.6.1
psycopg2~=2.9.3
requests~=2.28.1
python-dateutil~=2.8.2